refactor: Update project files
* Drop shell support from tools
  * remove `use_shell` and `shell_name` fields from tool
  * remove all related shell quote logic
* Remove task exec delay support
