+++ pkg/dukkha/context_exec_test.go
@@ -9,7 +9,6 @@ import (
 func TestTaskExecOptions(t *testing.T) {
 	const (
 		shellName = "test-shell"
-		useShell  = true
 	)
 
 	opts := CreateTaskExecOptions(1, 10).(*taskExecOpts)
@@ -18,7 +17,7 @@ func TestTaskExecOptions(t *testing.T) {
 	assert.Equal(t, opts.total, 10)
 
 	for i := 0; i < opts.total; i++ {
-		mOpts := opts.NextMatrixExecOptions(useShell, shellName).(*taskMatrixExecOpts)
+		mOpts := opts.NextMatrixExecOptions().(*taskMatrixExecOpts)
 		assert.Equal(t, opts.id, 1)
 		assert.Equal(t, opts.seq, i)
 		assert.Equal(t, opts.total, 10)
@@ -31,12 +30,6 @@ func TestTaskExecOptions(t *testing.T) {
 
 		assert.Equal(t, mOpts.total, mOpts.Total())
 		assert.Equal(t, mOpts.Total(), opts.total)
-
-		assert.Equal(t, mOpts.shellName, mOpts.ShellName())
-		assert.Equal(t, mOpts.ShellName(), shellName)
-
-		assert.Equal(t, mOpts.useShell, mOpts.UseShell())
-		assert.Equal(t, mOpts.UseShell(), useShell)
 
 		assert.Equal(t, i == opts.total-1, mOpts.IsLast())
 	}
