+++ pkg/dukkha/test/exec_opts.go
@@ -4,5 +4,5 @@ import "arhat.dev/dukkha/pkg/dukkha"
 
 func CreateTaskMatrixExecOptions() dukkha.TaskMatrixExecOptions {
 	opts := dukkha.CreateTaskExecOptions(1, 1)
-	return opts.NextMatrixExecOptions(false, "")
+	return opts.NextMatrixExecOptions()
 }
